requirements.txt
openai>=1.30.0
boto3>=1.34.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
